backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
python-multipart>=0.0.9
Pillow>=10.3.0
python-dotenv>=1.0.1
slowapi>=0.1.9
qrcode[pil]>=7.4.2
azure-cosmos>=4.5.0
msal>=1.24.0
httpx>=0.24.0
pytz>=2024.1
azure-storage-blob>=12.19.0
